abstract out flowNotebook
