+++ src/index.js
@@ -18,7 +18,6 @@ import { data } from './data/data';
 import { async } from './async/async';
 import { objectBrowser } from './objectBrowser/objectBrowser';
 import { help } from './help/help';
-import { notebook } from './notebook/notebook';
 import { failure } from './failure/failure';
 import { clipboard } from './clipboard/clipboard';
 import { about } from './about/about';
@@ -38,7 +37,6 @@ import { flow } from './flow/flow';
   clipboard();
   failure();
   help();
-  notebook();
   objectBrowser();
   async();
   data();
